api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
python-dotenv
google-genai
colorama
